+++ src/Providers/SMTP/Mailer.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace WPMailSMTP\Providers\SMTP;
+
+use WPMailSMTP\Providers\MailerAbstract;
+
+/**
+ * Class Mailer inherits everything from parent abstract class.
+ * This file is required for a proper work of Loader and \ReflectionClass.
+ *
+ * @package WPMailSMTP\Providers\SMTP
+ */
+class Mailer extends MailerAbstract {
+
+}