+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.0] - 2017-12-21
+### Fixed:
+- Decrease the factual minimum WordPress version from 3.9 to 3.6.
+
+### Changed: 
+- Improve debug output for all mail providers. 
+
 ## [1.1.0] - 2017-12-18
 ### Added: 
 - New option "Auto TLS" for SMTP mailer. Default is enabled. Migration routine for all sites.